Revert "Modular version"
