--- applications/CMakeLists.txt
@@ -1,18 +1,0 @@
-PROJECT(applications)
-
-INCLUDE_DIRECTORIES(
-    include
-    ${CMAKE_SOURCE_DIR}/include
-    )
-
-SET(SOURCE_FILES
-    main.cc
-    )
-
-FOREACH(sourcefile ${SOURCE_FILES})
-    # string replace: cut off .cc from files
-    STRING(REPLACE ".cc" "" testname ${sourcefile} )
-    ADD_EXECUTABLE( ${testname} ${sourcefile})
-    DEAL_II_SETUP_TARGET(${testname})
-    TARGET_LINK_LIBRARIES(${testname} rotatingMHD)
-ENDFOREACH( sourcefile ${APP_SOURCES} )
