+++ src/mod_step-35/parameter-file.prm
@@ -1,0 +1,42 @@
+# First a global definition
+# the type of method we want to use
+set Method_Form = rotational
+
+subsection Physical data
+  # In this subsection we declare the physical data
+  # The initial and final time, and the Reynolds number
+  set initial_time = 0.
+  set final_time   = 10.
+  set Reynolds     = 100
+end
+
+subsection Time step data
+  # In this subsection we declare the data that is to be used for time discretization,
+  # i.e. the time step dt
+  set dt = 5e-3
+end
+
+subsection Space discretization
+  # In this subsection we declare the data that is relevant to the space discretization
+  # we set the number of global refines the triangulation must have
+  # and the degree k of the pair Q_(k+1)--Q_k of velocity--pressure finite element spaces
+  set n_of_refines = 3
+  set pressure_fe_degree = 1
+end
+
+subsection Data solve velocity
+  # In this section we declare the parameters that are going to control the solution process
+  # for the velocity.
+  set max_iterations = 1000  # maximal number of iterations that GMRES must make
+  set eps            = 1e-8  # stopping criterion
+  set Krylov_size    = 30    # size of the Krylov subspace to be used in GMRES
+  set off_diagonals  = 70    # number of off diagonals that ILU must compute
+  set diag_strength  = 0.1  # diagonal strengthening value
+  set update_prec    = 10    # this number indicates how often the preconditioner must be updated
+end
+
+#The output frequency
+set output_interval = 50
+
+#Finally we set the verbosity level
+set verbose = false
