+++ src/mod_step-35/parameter_file.prm
@@ -7,7 +7,7 @@ end
 subsection Time discretization parameters
   set dt                              = 5e-3
   set t_0                             = 0.
-  set T                               = 5e-2
+  set T                               = 10.
 end
 
 subsection Space discretization parameters
@@ -25,5 +25,4 @@ subsection Parameters of the diffusion-step solver
 end
 
 set flag_verbose_output               = false
-set flag_adaptive_time_step           = false
-set output_interval                   = 10
+set output_interval                   = 50
